double

+++ oklab.cpp
@@ -1,36 +1,36 @@
-struct Lab {float L; float a; float b;};
-struct RGB {float r; float g; float b;};
+struct Lab {double L; double a; double b;};
+struct RGB {double r; double g; double b;};
 
 Lab linear_srgb_to_oklab(RGB c)
 {
-    float l = 0.4122214708f * c.r + 0.5363325363f * c.g + 0.0514459929f * c.b;
-	float m = 0.2119034982f * c.r + 0.6806995451f * c.g + 0.1073969566f * c.b;
-	float s = 0.0883024619f * c.r + 0.2817188376f * c.g + 0.6299787005f * c.b;
+    double l = 0.4122214708 * c.r + 0.5363325363 * c.g + 0.0514459929 * c.b;
+	double m = 0.2119034982 * c.r + 0.6806995451 * c.g + 0.1073969566 * c.b;
+	double s = 0.0883024619 * c.r + 0.2817188376 * c.g + 0.6299787005 * c.b;
 
-    float l_ = cbrtf(l);
-    float m_ = cbrtf(m);
-    float s_ = cbrtf(s);
+    double l_ = cbrt(l);
+    double m_ = cbrt(m);
+    double s_ = cbrt(s);
 
     return {
-        0.2104542553f*l_ + 0.7936177850f*m_ - 0.0040720468f*s_,
-        1.9779984951f*l_ - 2.4285922050f*m_ + 0.4505937099f*s_,
-        0.0259040371f*l_ + 0.7827717662f*m_ - 0.8086757660f*s_,
+        0.2104542553*l_ + 0.7936177850*m_ - 0.0040720468*s_,
+        1.9779984951*l_ - 2.4285922050*m_ + 0.4505937099*s_,
+        0.0259040371*l_ + 0.7827717662*m_ - 0.8086757660*s_,
     };
 }
 
 RGB oklab_to_linear_srgb(Lab c)
 {
-    float l_ = c.L + 0.3963377774f * c.a + 0.2158037573f * c.b;
-    float m_ = c.L - 0.1055613458f * c.a - 0.0638541728f * c.b;
-    float s_ = c.L - 0.0894841775f * c.a - 1.2914855480f * c.b;
+    double l_ = c.L + 0.3963377774 * c.a + 0.2158037573 * c.b;
+    double m_ = c.L - 0.1055613458 * c.a - 0.0638541728 * c.b;
+    double s_ = c.L - 0.0894841775 * c.a - 1.2914855480 * c.b;
 
-    float l = l_*l_*l_;
-    float m = m_*m_*m_;
-    float s = s_*s_*s_;
+    double l = l_*l_*l_;
+    double m = m_*m_*m_;
+    double s = s_*s_*s_;
 
     return {
-		+4.0767416621f * l - 3.3077115913f * m + 0.2309699292f * s,
-		-1.2684380046f * l + 2.6097574011f * m - 0.3413193965f * s,
-		-0.0041960863f * l - 0.7034186147f * m + 1.7076147010f * s,
+		+4.0767416621 * l - 3.3077115913 * m + 0.2309699292 * s,
+		-1.2684380046 * l + 2.6097574011 * m - 0.3413193965 * s,
+		-0.0041960863 * l - 0.7034186147 * m + 1.7076147010 * s,
     };
 }
